+++ web/packages/pyxis-components/src/public/organisms/EthosStrip/EthosStrip.tsx
@@ -1,5 +1,5 @@
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import './EthosStrip.css';
 
 export type EthosStripProps = { className?: string };