+++ web/packages/pyxis-components/src/public/organisms/MailingListCTA/MailingListCTA.stories.tsx
@@ -1,7 +1,7 @@
 import type { CSSProperties } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { MailingListCTA } from './MailingListCTA';
-const meta: Meta<typeof MailingListCTA> = { title: 'Public/Organisms/MailingListCTA', component: MailingListCTA, tags: ['autodocs'] };
+const meta: Meta<typeof MailingListCTA> = { title: 'Public Site/Components/Organisms/MailingListCTA', component: MailingListCTA, tags: ['autodocs'] };
 export default meta;
 type Story = StoryObj<typeof MailingListCTA>;
 export const Default: Story = { args: {} };