+++ web/packages/pyxis-components/src/public/organisms/MailingListCTA/MailingListCTA.tsx
@@ -1,7 +1,7 @@
 import type { FormEvent } from 'react';
 import { useState } from 'react';
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import './MailingListCTA.css';
 
 export type MailingListCTAProps = {