+++ web/packages/pyxis-components/src/public/organisms/PubFooter/PubFooter.stories.tsx
@@ -1,7 +1,7 @@
 import type { CSSProperties } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { PubFooter } from './PubFooter';
-const meta: Meta<typeof PubFooter> = { title: 'Public/Organisms/PubFooter', component: PubFooter, tags: ['autodocs'] };
+const meta: Meta<typeof PubFooter> = { title: 'Public Site/Components/Organisms/PubFooter', component: PubFooter, tags: ['autodocs'] };
 export default meta;
 type Story = StoryObj<typeof PubFooter>;
 export const Default: Story = {};