+++ web/packages/pyxis-components/src/public/organisms/PubHero/PubHero.tsx
@@ -1,5 +1,5 @@
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import type { Show } from 'pyxis-types';
 import './PubHero.css';